deal-sniper/requirements.txt
telethon
requests
beautifulsoup4
python-dotenv
pytest
